WIP: product APIs

+++ src/main/java/org/sir/stripeintegration/core/application/interfaces/service/ICustomerService.java
@@ -9,7 +9,7 @@ import reactor.core.publisher.Mono;
 public interface ICustomerService {
     Mono<CustomerDto> getCustomer(String id);
 
-    Flux<CustomerDto> getAllCustomer(Integer limit, String startingAfter, String endingBefore);
+    Flux<CustomerDto> getAllCustomer(Long limit, String startingAfter, String endingBefore);
 
     Mono<CustomerDto> addCustomer(CustomerCreateRequestDto requestDto);
