+++ src/main/java/org/sir/stripeintegration/infrastructure/service/PaymentMethodService.java
@@ -40,7 +40,7 @@ public class PaymentMethodService implements IPaymentMethodService {
 
     @Override
     public Flux<PaymentMethodDto> getCustomerAllPaymentMethod(
-            String customerId, Integer limit, String startingAfter, String endingBefore) {
+            String customerId, Long limit, String startingAfter, String endingBefore) {
         return customerRepository.findById(customerId)
                 .switchIfEmpty(Mono.error(new CustomException(ErrorMessage.CUSTOMER_NOT_FOUND.getMessage())))
                 .map(customerEntity -> stripeRootService.getCustomerAllPaymentMethods(
